+++ util/options_helper.cc
@@ -6,7 +6,10 @@
 #include <cassert>
 #include <cctype>
 #include <unordered_set>
+#include "rocksdb/cache.h"
+#include "rocksdb/filter_policy.h"
 #include "rocksdb/options.h"
+#include "rocksdb/table.h"
 #include "rocksdb/utilities/convenience.h"
 #include "util/options_helper.h"
 
@@ -29,9 +32,31 @@ CompressionType ParseCompressionType(const std::string& type) {
   } else if (type == "kLZ4HCCompression") {
     return kLZ4HCCompression;
   } else {
-    throw "unknown compression type: " + type;
+    throw std::invalid_argument("Unknown compression type: " + type);
   }
   return kNoCompression;
+}
+
+BlockBasedTableOptions::IndexType ParseBlockBasedTableIndexType(
+    const std::string& type) {
+  if (type == "kBinarySearch") {
+    return BlockBasedTableOptions::kBinarySearch;
+  } else if (type == "kHashSearch") {
+    return BlockBasedTableOptions::kHashSearch;
+  }
+  throw std::invalid_argument("Unknown index type: " + type);
+}
+
+ChecksumType ParseBlockBasedTableChecksumType(
+    const std::string& type) {
+  if (type == "kNoChecksum") {
+    return kNoChecksum;
+  } else if (type == "kCRC32c") {
+    return kCRC32c;
+  } else if (type == "kxxHash") {
+    return kxxHash;
+  }
+  throw std::invalid_argument("Unknown checksum type: " + type);
 }
 
 bool ParseBoolean(const std::string& type, const std::string& value) {
@@ -39,9 +64,8 @@ bool ParseBoolean(const std::string& type, const std::string& value) {
     return true;
   } else if (value == "false" || value == "0") {
     return false;
-  } else {
-    throw type;
   }
+  throw std::invalid_argument(type);
 }
 
 uint64_t ParseUint64(const std::string& value) {
@@ -105,7 +129,7 @@ CompactionStyle ParseCompactionStyle(const std::string& type) {
   } else if (type == "kCompactionStyleFIFO") {
     return kCompactionStyleFIFO;
   } else {
-    throw "unknown compaction style: " + type;
+    throw std::invalid_argument("unknown compaction style: " + type);
   }
   return kCompactionStyleLevel;
 }
@@ -172,7 +196,7 @@ bool ParseCompactionOptions(const std::string& name, const std::string& value,
     new_options->max_bytes_for_level_multiplier_additional.clear();
     size_t start = 0;
     while (true) {
-      size_t end = value.find_first_of(':', start);
+      size_t end = value.find(':', start);
       if (end == std::string::npos) {
         new_options->max_bytes_for_level_multiplier_additional.push_back(
             ParseInt(value.substr(start)));
@@ -210,8 +234,8 @@ Status GetMutableOptionsFromStrings(
     MutableCFOptions* new_options) {
   assert(new_options);
   *new_options = base_options;
-  try {
-    for (const auto& o : options_map) {
+  for (const auto& o : options_map) {
+    try {
       if (ParseMemtableOptions(o.first, o.second, new_options)) {
       } else if (ParseCompactionOptions(o.first, o.second, new_options)) {
       } else if (ParseMiscOptions(o.first, o.second, new_options)) {
@@ -219,9 +243,10 @@ Status GetMutableOptionsFromStrings(
         return Status::InvalidArgument(
             "unsupported dynamic option: " + o.first);
       }
+    } catch (std::exception& e) {
+      return Status::InvalidArgument("error parsing " + o.first + ":" +
+                                     std::string(e.what()));
     }
-  } catch (std::exception& e) {
-    return Status::InvalidArgument("error parsing " + std::string(e.what()));
   }
   return Status::OK();
 }
@@ -243,38 +268,165 @@ std::string trim(const std::string& str) {
   return std::string();
 }
 
-bool StringToMap(const std::string& opts_str,
-                 std::unordered_map<std::string, std::string>* opts_map) {
+}  // anonymous namespace
+
+Status StringToMap(const std::string& opts_str,
+                   std::unordered_map<std::string, std::string>* opts_map) {
   assert(opts_map);
   // Example:
-  //   opts_str = "write_buffer_size=1024;max_write_buffer_number=2"
+  //   opts_str = "write_buffer_size=1024;max_write_buffer_number=2;"
+  //              "nested_opt={opt1=1;opt2=2};max_bytes_for_level_base=100"
   size_t pos = 0;
-
   std::string opts = trim(opts_str);
   while (pos < opts.size()) {
     size_t eq_pos = opts.find('=', pos);
     if (eq_pos == std::string::npos) {
-      return false;
+      return Status::InvalidArgument("Mismatched key value pair, '=' expected");
     }
     std::string key = trim(opts.substr(pos, eq_pos - pos));
-
-    size_t sc_pos = opts.find(';', eq_pos + 1);
-    if (sc_pos == std::string::npos) {
-      (*opts_map)[key] = trim(opts.substr(eq_pos + 1));
-      // It either ends with a trailing semi-colon or the last key-value pair
-      break;
-    } else {
-      (*opts_map)[key] = trim(opts.substr(eq_pos + 1, sc_pos - eq_pos - 1));
+    if (key.empty()) {
+      return Status::InvalidArgument("Empty key found");
     }
-    pos = sc_pos + 1;
+
+    // skip space after '=' and look for '{' for possible nested options
+    pos = eq_pos + 1;
+    while (pos < opts.size() && isspace(opts[pos])) {
+      ++pos;
+    }
+    // Empty value at the end
+    if (pos >= opts.size()) {
+      (*opts_map)[key] = "";
+      break;
+    }
+    if (opts[pos] == '{') {
+      int count = 1;
+      size_t brace_pos = pos + 1;
+      while (brace_pos < opts.size()) {
+        if (opts[brace_pos] == '{') {
+          ++count;
+        } else if (opts[brace_pos] == '}') {
+          --count;
+          if (count == 0) {
+            break;
+          }
+        }
+        ++brace_pos;
+      }
+      // found the matching closing brace
+      if (count == 0) {
+        (*opts_map)[key] = trim(opts.substr(pos + 1, brace_pos - pos - 1));
+        // skip all whitespace and move to the next ';'
+        // brace_pos points to the next position after the matching '}'
+        pos = brace_pos + 1;
+        while (pos < opts.size() && isspace(opts[pos])) {
+          ++pos;
+        }
+        if (pos < opts.size() && opts[pos] != ';') {
+          return Status::InvalidArgument(
+              "Unexpected chars after nested options");
+        }
+        ++pos;
+      } else {
+        return Status::InvalidArgument(
+            "Mismatched curly braces for nested options");
+      }
+    } else {
+      size_t sc_pos = opts.find(';', pos);
+      if (sc_pos == std::string::npos) {
+        (*opts_map)[key] = trim(opts.substr(pos));
+        // It either ends with a trailing semi-colon or the last key-value pair
+        break;
+      } else {
+        (*opts_map)[key] = trim(opts.substr(pos, sc_pos - pos));
+      }
+      pos = sc_pos + 1;
+    }
   }
 
-  return true;
+  return Status::OK();
 }
 
-}  // anonymous namespace
 
-bool GetColumnFamilyOptionsFromMap(
+Status GetBlockBasedTableOptionsFromMap(
+    const BlockBasedTableOptions& table_options,
+    const std::unordered_map<std::string, std::string>& opts_map,
+    BlockBasedTableOptions* new_table_options) {
+
+  assert(new_table_options);
+  *new_table_options = table_options;
+  for (const auto& o : opts_map) {
+    try {
+      if (o.first == "cache_index_and_filter_blocks") {
+        new_table_options->cache_index_and_filter_blocks =
+          ParseBoolean(o.first, o.second);
+      } else if (o.first == "index_type") {
+        new_table_options->index_type = ParseBlockBasedTableIndexType(o.second);
+      } else if (o.first == "hash_index_allow_collision") {
+        new_table_options->hash_index_allow_collision =
+          ParseBoolean(o.first, o.second);
+      } else if (o.first == "checksum") {
+        new_table_options->checksum =
+          ParseBlockBasedTableChecksumType(o.second);
+      } else if (o.first == "no_block_cache") {
+        new_table_options->no_block_cache = ParseBoolean(o.first, o.second);
+      } else if (o.first == "block_cache") {
+        new_table_options->block_cache = NewLRUCache(ParseSizeT(o.second));
+      } else if (o.first == "block_cache_compressed") {
+        new_table_options->block_cache_compressed =
+          NewLRUCache(ParseSizeT(o.second));
+      } else if (o.first == "block_size") {
+        new_table_options->block_size = ParseSizeT(o.second);
+      } else if (o.first == "block_size_deviation") {
+        new_table_options->block_size_deviation = ParseInt(o.second);
+      } else if (o.first == "block_restart_interval") {
+        new_table_options->block_restart_interval = ParseInt(o.second);
+      } else if (o.first == "filter_policy") {
+        // Expect the following format
+        // bloomfilter:int:bool
+        const std::string kName = "bloomfilter:";
+        if (o.second.compare(0, kName.size(), kName) != 0) {
+          return Status::InvalidArgument("Invalid filter policy name");
+        }
+        size_t pos = o.second.find(':', kName.size());
+        if (pos == std::string::npos) {
+          return Status::InvalidArgument("Invalid filter policy config, "
+                                         "missing bits_per_key");
+        }
+        int bits_per_key = ParseInt(
+            trim(o.second.substr(kName.size(), pos - kName.size())));
+        bool use_block_based_builder =
+          ParseBoolean("use_block_based_builder",
+                       trim(o.second.substr(pos + 1)));
+        new_table_options->filter_policy.reset(
+            NewBloomFilterPolicy(bits_per_key, use_block_based_builder));
+      } else if (o.first == "whole_key_filtering") {
+        new_table_options->whole_key_filtering =
+          ParseBoolean(o.first, o.second);
+      } else {
+        return Status::InvalidArgument("Unrecognized option: " + o.first);
+      }
+    } catch (std::exception& e) {
+      return Status::InvalidArgument("error parsing " + o.first + ":" +
+                                     std::string(e.what()));
+    }
+  }
+  return Status::OK();
+}
+
+Status GetBlockBasedTableOptionsFromString(
+    const BlockBasedTableOptions& table_options,
+    const std::string& opts_str,
+    BlockBasedTableOptions* new_table_options) {
+  std::unordered_map<std::string, std::string> opts_map;
+  Status s = StringToMap(opts_str, &opts_map);
+  if (!s.ok()) {
+    return s;
+  }
+  return GetBlockBasedTableOptionsFromMap(table_options, opts_map,
+                                          new_table_options);
+}
+
+Status GetColumnFamilyOptionsFromMap(
     const ColumnFamilyOptions& base_options,
     const std::unordered_map<std::string, std::string>& opts_map,
     ColumnFamilyOptions* new_options) {
@@ -285,6 +437,15 @@ bool GetColumnFamilyOptionsFromMap(
       if (ParseMemtableOptions(o.first, o.second, new_options)) {
       } else if (ParseCompactionOptions(o.first, o.second, new_options)) {
       } else if (ParseMiscOptions(o.first, o.second, new_options)) {
+      } else if (o.first == "block_based_table_factory") {
+        // Nested options
+        BlockBasedTableOptions table_opt;
+        Status table_opt_s = GetBlockBasedTableOptionsFromString(
+            BlockBasedTableOptions(), o.second, &table_opt);
+        if (!table_opt_s.ok()) {
+          return table_opt_s;
+        }
+        new_options->table_factory.reset(NewBlockBasedTableFactory(table_opt));
       } else if (o.first == "min_write_buffer_number_to_merge") {
         new_options->min_write_buffer_number_to_merge = ParseInt(o.second);
       } else if (o.first == "compression") {
@@ -293,7 +454,7 @@ bool GetColumnFamilyOptionsFromMap(
         new_options->compression_per_level.clear();
         size_t start = 0;
         while (true) {
-          size_t end = o.second.find_first_of(':', start);
+          size_t end = o.second.find(':', start);
           if (end == std::string::npos) {
             new_options->compression_per_level.push_back(
                 ParseCompressionType(o.second.substr(start)));
@@ -306,22 +467,25 @@ bool GetColumnFamilyOptionsFromMap(
         }
       } else if (o.first == "compression_opts") {
         size_t start = 0;
-        size_t end = o.second.find_first_of(':');
+        size_t end = o.second.find(':');
         if (end == std::string::npos) {
-          throw o.first;
+          return Status::InvalidArgument("invalid config value for: "
+                                         + o.first);
         }
         new_options->compression_opts.window_bits =
             ParseInt(o.second.substr(start, end - start));
         start = end + 1;
-        end = o.second.find_first_of(':', start);
+        end = o.second.find(':', start);
         if (end == std::string::npos) {
-          throw o.first;
+          return Status::InvalidArgument("invalid config value for: "
+                                         + o.first);
         }
         new_options->compression_opts.level =
             ParseInt(o.second.substr(start, end - start));
         start = end + 1;
         if (start >= o.second.size()) {
-          throw o.first;
+          return Status::InvalidArgument("invalid config value for: "
+                                         + o.first);
         }
         new_options->compression_opts.strategy =
             ParseInt(o.second.substr(start, o.second.size() - start));
@@ -334,7 +498,7 @@ bool GetColumnFamilyOptionsFromMap(
         new_options->compaction_style = ParseCompactionStyle(o.second);
       } else if (o.first == "compaction_options_universal") {
         // TODO(ljin): add support
-        throw o.first;
+        return Status::NotSupported("Not supported: " + o.first);
       } else if (o.first == "compaction_options_fifo") {
         new_options->compaction_options_fifo.max_table_files_size
           = ParseUint64(o.second);
@@ -345,27 +509,29 @@ bool GetColumnFamilyOptionsFromMap(
       } else if (o.first == "inplace_update_support") {
         new_options->inplace_update_support = ParseBoolean(o.first, o.second);
       } else {
-        return false;
+        return Status::InvalidArgument("Unrecognized option: " + o.first);
       }
-    } catch (std::exception) {
-      return false;
+    } catch (std::exception& e) {
+      return Status::InvalidArgument("error parsing " + o.first + ":" +
+                                     std::string(e.what()));
     }
   }
-  return true;
+  return Status::OK();
 }
 
-bool GetColumnFamilyOptionsFromString(
+Status GetColumnFamilyOptionsFromString(
     const ColumnFamilyOptions& base_options,
     const std::string& opts_str,
     ColumnFamilyOptions* new_options) {
   std::unordered_map<std::string, std::string> opts_map;
-  if (!StringToMap(opts_str, &opts_map)) {
-    return false;
+  Status s = StringToMap(opts_str, &opts_map);
+  if (!s.ok()) {
+    return s;
   }
   return GetColumnFamilyOptionsFromMap(base_options, opts_map, new_options);
 }
 
-bool GetDBOptionsFromMap(
+Status GetDBOptionsFromMap(
     const DBOptions& base_options,
     const std::unordered_map<std::string, std::string>& opts_map,
     DBOptions* new_options) {
@@ -392,7 +558,7 @@ bool GetDBOptionsFromMap(
         new_options->use_fsync = ParseBoolean(o.first, o.second);
       } else if (o.first == "db_paths") {
         // TODO(ljin): add support
-        throw o.first;
+        return Status::NotSupported("Not supported: " + o.first);
       } else if (o.first == "db_log_dir") {
         new_options->db_log_dir = o.second;
       } else if (o.first == "wal_dir") {
@@ -444,22 +610,24 @@ bool GetDBOptionsFromMap(
       } else if (o.first == "bytes_per_sync") {
         new_options->bytes_per_sync = ParseUint64(o.second);
       } else {
-        return false;
+        return Status::InvalidArgument("Unrecognized option: " + o.first);
       }
-    } catch (std::exception) {
-      return false;
+    } catch (std::exception& e) {
+      return Status::InvalidArgument("error parsing " + o.first + ":" +
+                                     std::string(e.what()));
     }
   }
-  return true;
+  return Status::OK();
 }
 
-bool GetDBOptionsFromString(
+Status GetDBOptionsFromString(
     const DBOptions& base_options,
     const std::string& opts_str,
     DBOptions* new_options) {
   std::unordered_map<std::string, std::string> opts_map;
-  if (!StringToMap(opts_str, &opts_map)) {
-    return false;
+  Status s = StringToMap(opts_str, &opts_map);
+  if (!s.ok()) {
+    return s;
   }
   return GetDBOptionsFromMap(base_options, opts_map, new_options);
 }